Rename tokenomics component and randomize chart colors
