+++ public/locales/en/home.json
@@ -9,7 +9,6 @@
   "today": "Today",
   "lastMonth": "Last Month",
   "consensus": "Consensus",
-  "tokenomics": "Tokenomics",
   "bondedPercent": "Bonded\n{{percent}}",
   "unbondedPercent": "Unbonded\n{{percent}}",
   "unbondingPercent": "Unbonding\n{{percent}}",
