+++ src/screens/home/components/index.ts
@@ -1,7 +1,7 @@
 import DataBlocks from './data_blocks';
 import OnlineVotingPower from './online_voting_power';
 import Consensus from './consensus';
-import Tokenomics from './tokenomics';
+import ValidatorsChart from './validators_chart';
 import Blocks from './blocks';
 import Transactions from './transactions';
 
@@ -9,7 +9,7 @@ export {
   DataBlocks,
   OnlineVotingPower,
   Consensus,
-  Tokenomics,
+  ValidatorsChart,
   Blocks,
   Transactions,
 };
