+++ src/screens/home/components/validators_chart/index.test.tsx
