+++ src/screens/home/components/validators_chart/styles.tsx
